Solve the problem of display
