+++ templates/base.html
@@ -1,0 +1,68 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="../static/HomePage-css.css" />
+    <link rel="stylesheet" href="../static/profile_styles.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+    <script type="module" src="../static/HomePage_t.js"></script>
+    <script type="module" src="../static/profile_test.js"></script>
+
+    <title>Document</title>
+</head>
+
+<body>
+    <div id="navigation-area">
+        <div id="logo">
+            <img src="../static/pic/HomePage-image/logo.png" alt="" />
+        </div>
+        <div id="navigation">
+            <ul>
+                <li><img src="../static/pic/HomePage-image/home.png" alt="" /><a href="/index">Home</a></li>
+                <li>
+                    <img src="../static/pic/HomePage-image/explore.png" alt="" />Explore
+                </li>
+                <li>
+                    <img src="../static/pic/HomePage-image/notification.png" alt="" />Notifications
+                </li>
+                <li>
+                    <img src="../static/pic/HomePage-image/message.png" alt="" />Messages
+                </li>
+                <li>
+                    <img src="../static/pic/HomePage-image/communities.png" alt="" />Communities
+                </li>
+                <li>
+                    <img src="../static/pic/HomePage-image/profile.png" alt="" /><a href="/profile">Profile</a>
+                </li>
+                <li><img src="../static/pic/HomePage-image/more.png" alt="" />More</li>
+            </ul>
+        </div>
+        <div id="post"><a href="" class="default-button">Post</a></div>
+        <div id="profile">
+            <a href="" class="default-button">
+                <div>
+                    <img class="avatarImg" src="" alt="" />
+                </div>
+                <div>
+                    <p class="currentUser"></p>
+                </div>
+                <div>
+                    <img src="../static/pic/HomePage-image/dotdotdot.png" alt="" />
+                </div>
+            </a>
+        </div>
+        <div><a href="/login">Login</a></div>
+        <div id="homepagelogout"><a href="/logout">Logout</a></div>
+    </div>
+
+    {% block content %}
+    {% endblock %}
+
+    {% block sideArea %}
+    {% endblock %}
+</body>
+
+</html>
