+++ templates/index.html
@@ -31,7 +31,7 @@
                     <img src="../static/pic/HomePage-image/communities.png" alt="" />Communities
                 </li>
                 <li>
-                    <img src="../static/pic/HomePage-image/profile.png" alt="" />Profile
+                    <img src="../static/pic/HomePage-image/profile.png" alt="" /><a href="/profile">Profile</a>
                 </li>
                 <li><img src="../static/pic/HomePage-image/more.png" alt="" />More</li>
             </ul>
